+++ src/monitor.rs
@@ -79,7 +79,7 @@ impl Monitor {
         self.impl_monitor.capture_image()
     }
 
-    pub fn screenshot(&self, x: i32, y: i32, width: i32, height: i32) -> XCapResult<RgbaImage> {
+    pub fn screenshot(&self, x: f64, y: f64, width: f64, height: f64) -> XCapResult<RgbaImage> {
         self.impl_monitor.screenshot(x, y, width, height)
     }
